chore: moved sqlalchemy manager to own file
